+++ mall-manage/mall-manage-service/src/main/java/com/mall/manage/service/ItemCatService.java
@@ -15,5 +15,5 @@ public interface ItemCatService {
 	 * @param parentId
 	 * @return
 	 */
-	public List<ItemCat> getListByParentId(Long parentId);
+	public List<ItemCat> getListByParentId(int parentId);
 }
